+++ app/src/main/java/com/cajetan/youtubeplus/MainActivity.kt
@@ -33,7 +33,7 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         val darkMode = PreferenceManager.getDefaultSharedPreferences(this)
-                .getBoolean("dark_mode", false)
+                .getBoolean(getString(R.string.dark_mode_key), false)
         setTheme(if (darkMode) R.style.AppThemeDark else R.style.AppThemeLight)
 
         super.onCreate(savedInstanceState)